Separate concerns between serializer and dataStore better
There was an issue where I was trying to decide at what step the
dataStore should be attached to the model. I was trying to do it in the
dataStore, but that required the store to be aware of what the model or
collection looked like, so there was some coupling. Now the dataStore
just hands a reference of itself to the serializer, and the serializer
bakes it in