Make reference attaching more concise.
Attaching references to models is really only the concern of the client
library. So long as the model's constructor has a facility to pass in
extra attrs, that should be enough for any model. And so long as the
serializer can handle passing that data through as not a main part of
the attr set, should be good?

This also puts the responsibility of wiring all the things together
solely on the client, which is the thing that wires all the things
together.