Helpers for building graph queries.
This sort of makes the process of writing graph queries easier. The
syntax is more verbose than I'd like, so I'm open to suggestions on a
better approach.

- A guarantee that if you're code works, you're asking graph API data
that exists.
- Don't know what fields exist on a resource? Don't know resource
relationships? Look at the schema module in the dev console and
introspect all the things, OR use helpers to inject all the things into
your requests.
- Paginated requests. You don't want to have to write your own
nodes/edges/pageInfo fetching code, so this wraps that up.