Graph Query Helpers (#181)
* Helpers for building graph queries.

This sort of makes the process of writing graph queries easier. The
syntax is more verbose than I'd like, so I'm open to suggestions on a
better approach.

- A guarantee that if you're code works, you're asking graph API data
that exists.
- Don't know what fields exist on a resource? Don't know resource
relationships? Look at the schema module in the dev console and
introspect all the things, OR use helpers to inject all the things into
your requests.
- Paginated requests. You don't want to have to write your own
nodes/edges/pageInfo fetching code, so this wraps that up.

Use `join` for field joints as well.

* Protect the schema from accidental mutations

* Add the base query helper

Refactor the query helper into a higher level wrapper so you don't need
to import the schema yourself.

Normalize join on args.

Don't join empty strings.

* Add leaky npm dep that isn't included in fs-tree-diff