Remove tags field from Product fragment #597
The default ProductFragment provided by this SDK comes with a `tags` field,
but since the permission to read product tags is disabled by default in the
Storefront API settings, an error would occurr with the message "access denied"
when using `client.product.fetchAll()`.

This PR fixes the issue by removing the `tags` field from the Product fragment.