Small discount error fixups
As a sanity check during the development phase, we were throwing
an error if we were aware of a discount code but it wasn't appearing
in the output of our discount mapping. Upon investigating further, we
determined that this is expected behaviour for any discount codes that
are applied but have `applicable: false`. Therefore, we are updating
our sanity check to only throw an error if the discount code IS applied
but somehow didn't get mapped.

These sanity checks that throw errors will be removed before release.