Return currency amounts as strings, matching JS Buy SDK v2
A shop testing our v3 RC reported that the subtotalPrice field was always being
rendered in their shop  as $0.

We realized this is because that shop expected the currency amounts to be strings
(like what the SF API/JS Buy SDK v2 return), but we were returning some currency
amounts as numbers.

This change ensures that all currency fields we return should be of type string
rather than type number, which should resolve the bug for that shop.

Normalize all currency fields in the checkout payload to contain the same number
of decimal places that would be returned by the storefront API.

In the storefront API, currency amounts are returned as a string that contains
1 decimal place (if the 2nd decimal place is 0) or else 2 decimal places.

In our mapping functions, we are typically converting to strings with 2 decimal
places. In case any clients of the JS Buy SDK are relying only a single decimal
place being returned in some cases, we want to normalize the decimal places.