Gracefully abort discount mapping if errors occur
If discount mapping fails, we don't want the errors to bubble up to the consumer and break
all operations they attempt. Instead, if discount mapping fails, we will abort all discount
mapping and map the cart API payload as if there are no discounts at all.\

In the near future, it would be helpful to integrate with something like Bugsnag so that we
can get visibility on these errors and the original cart API payload that we attempted to map
when these errors occurred, without consumers needing to manually send us this information.