Fix bug with adding multiple discount codes when not all codes are applicable
A discount code can be added to a cart but might not be `applicable`. In that case, the code
will appear in the `discountCodes` array, but there won't be any discount allocations.

Since we were using discount allocations to determine which discount codes existed, we would
end up inadvertently removing discount codes that were added but not `applicable`.